Moves benchmarking to new file
